adjusted tests
